# src/main/resources/static/openapi.yml
openapi: 3.0.0
info:
  title: "jeju API"
  description: "jeju API"
  version: "1.0.0"
servers:
  - url: "https://jeju"
...
paths:
...
  /api/v1/auth/user/{userId}:
    patch:
      summary: "PATCH api/v1/auth/user/{userId}"
      tags:
        - User
      operationId: "updatePassword"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInfoUpdateRequestDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserInfoResponseDto"
...
  /api/v1/auth/sms/check:
    post:
      tags:
        - Authentication
      summary: "POST api/v1/auth/sms/check"
      operationId: "checkUserUsingVerificationCode"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SMSVerificationDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /api/v1/auth/sms/check/later:
    post:
      tags:
        - Authentication
      summary: "POST api/v1/auth/sms/check/later"
      operationId: "checkUserUsingVerificationCodeAfterSignup"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SMSVerificationDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SMSVerificationResultResponse"
  /api/v1/auth/sms/send:
    post:
      tags:
        - Authentication
      summary: "POST api/v1/auth/sms/send"
      operationId: "sendSMSForVerification"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SMSMessageDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /api/v1/auth/login:
    post:
      tags:
        - User
      summary: "POST api/v1/auth/login"
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
...
  /api/v1/auth/signup:
    post:
      tags:
        - User
      summary: "POST api/v1/auth/signup"
      operationId: "signup"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
...
  /api/v1/auth/check/id:
    post:
      tags:
        - User
      summary: "POST api/v1/auth/check/id"
      operationId: "checkIdDuplicate"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequestDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /api/v1/auth/find/id:
    post:
      tags:
        - User
      summary: "POST api/v1/auth/find/id"
      operationId: "getId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindAuthByPhoneRequestDto"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FindAuthResponseDto"
  /api/v1/auth/authentication/check:
    post:
      tags:
        - Authentication
      summary: "POST api/v1/auth/authentication/check"
      operationId: "checkAuthenticationcode"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticationRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
...
components:
  schemas:
...
    UserInfoResponseDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        email:
          type: "string"
        nickname:
          type: "string"
        name:
          type: "string"
        numberTag:
          type: "integer"
          format: "int64"
        profileImageUrl:
          type: "string"
        userRole:
          type: "string"
          enum:
            - "USER"
            - "AUTHOR"
...
    SMSVerificationDTO:
      type: "object"
      properties:
        name:
          type: "string"
        phoneNumber:
          type: "string"
        verificationCode:
          type: "string"
    SMSMessageDTO:
      type: "object"
      properties:
        name:
          type: "string"
        phoneNumber:
          type: "string"
    LoginRequest:
      type: "object"
      properties:
        email:
          type: "string"
        password:
          type: "string"
    SignupRequest:
      type: "object"
      properties:
        email:
          type: "string"
        password:
          type: "string"
        nickname:
          type: "string"
        name:
          type: "string"
        phoneNumber:
          type: "string"
...
    AuthenticationRequest:
      type: "object"
      properties:
        email:
          type: "string"
        authCode:
          type: "string"
    EmailRequestDto:
      type: "object"
      properties:
        email:
          type: "string"
    FindAuthByPhoneRequestDto:
      type: "object"
      properties:
        name:
          type: "string"
        phoneNumber:
          type: "string"
    FindAuthResponseDto:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        email:
          type: "string"
        name:
          type: "string"
        createdAt:
          type: "string"
          format: "date-time"
    UserInfoUpdateRequestDto:
      type: "object"
      properties:
        password:
          type: "string"
        name:
          type: "string"
    SMSVerificationResultResponse:
      type: "object"
      properties:
        userId:
          type: "integer"
          format: "int64"